Update ItemUtils.cs
After actually testing it and making changes where necessary, this commit will very definitely fix the shops. Except for the stock that Pierre and Willy sell from what you directly sell to them. Possibly also the dish of the day. The Adventurer's Guild shop is coming in an update to another file.